Added Anna Videos + App.js for video render ONLY
These are all the Anna videos used for background. as a test render, the script App.js only uses Anna.mp4 so please disregard the other videos if you only want to see how it runs. this repo also serves as our backup repo so it makes sense to put them here